More project file fixup (#16)
* Use local nuget feed to reference built lottie package

* Fix non-x86 deployment in sample app, clean up local nuget feeds

* Cleaned up SimpleLottieSample.vcxproj some more

* Revert to using a project reference now that the nuget package has been tested, fix typo in readme

* Fix naming in nuget target files, fix float to double build warnings

+++ Samples/Islands/Lottie/LottieIsland/LottieContentIsland.cpp
@@ -77,12 +77,12 @@ namespace winrt::LottieIsland::implementation
         return m_progressPropertySet != nullptr;
     }
 
-    double LottieContentIsland::PlaybackRate() const
+    float LottieContentIsland::PlaybackRate() const
     {
         return m_playbackRate;
     }
 
-    void LottieContentIsland::PlaybackRate(double rate)
+    void LottieContentIsland::PlaybackRate(float rate)
     {
         m_playbackRate = rate;
         if (m_animationController != nullptr)
@@ -99,7 +99,7 @@ namespace winrt::LottieIsland::implementation
         }
     }
 
-    winrt::Windows::Foundation::IAsyncAction LottieContentIsland::PlayAsync(double fromProgress, double toProgress, bool looped)
+    winrt::Windows::Foundation::IAsyncAction LottieContentIsland::PlayAsync(float fromProgress, float toProgress, bool looped)
     {
         // Stop any existing animation
         StopAnimation();
@@ -123,7 +123,7 @@ namespace winrt::LottieIsland::implementation
         StopAnimation();
     }
 
-    void LottieContentIsland::StartAnimation(double fromProgress, double toProgress, bool loop)
+    void LottieContentIsland::StartAnimation(float fromProgress, float toProgress, bool loop)
     {
         if (m_animatedVisual == nullptr)
         {
